+++ unique_toolkit/unique_toolkit/language_model/infos.py
@@ -1,6 +1,6 @@
 from datetime import date
 from enum import StrEnum
-from typing import ClassVar, Optional
+from typing import ClassVar, Optional, Self
 
 from pydantic import BaseModel
 from typing_extensions import deprecated
@@ -11,13 +11,14 @@ from unique_toolkit.language_model.schemas import LanguageModelTokenLimits
 class LanguageModelName(StrEnum):
     AZURE_GPT_35_TURBO_0125 = "AZURE_GPT_35_TURBO_0125"
     AZURE_GPT_4_0613 = "AZURE_GPT_4_0613"
-    AZURE_GPT_4_TURBO_1106 = "AZURE_GPT_4_TURBO_1106"
-    AZURE_GPT_4_VISION_PREVIEW = "AZURE_GPT_4_VISION_PREVIEW"
     AZURE_GPT_4_32K_0613 = "AZURE_GPT_4_32K_0613"
     AZURE_GPT_4_TURBO_2024_0409 = "AZURE_GPT_4_TURBO_2024_0409"
     AZURE_GPT_4o_2024_0513 = "AZURE_GPT_4o_2024_0513"
     AZURE_GPT_4o_2024_0806 = "AZURE_GPT_4o_2024_0806"
     AZURE_GPT_4o_MINI_2024_0718 = "AZURE_GPT_4o_MINI_2024_0718"
+    AZURE_GPT_o1_2024_1217 = "AZURE_GPT_o1_2024_1217"
+    AZURE_GPT_o1_MINI_2024_0912 = "AZURE_GPT_o1_MINI_2024_0912"
+    AZURE_GPT_o3_MINI_2025_0131 = "AZURE_GPT_o3_MINI_2025_0131"
 
 
 class EncoderName(StrEnum):
@@ -25,15 +26,13 @@ class EncoderName(StrEnum):
     CL100K_BASE = "cl100k_base"
 
 
-def get_encoder_name(model_name: LanguageModelName) -> Optional[EncoderName]:
+def get_encoder_name(model_name: LanguageModelName) -> EncoderName:
     LMN = LanguageModelName
     match model_name:
         case LMN.AZURE_GPT_35_TURBO_0125:
             return EncoderName.CL100K_BASE
         case (
             LMN.AZURE_GPT_4_0613
-            | LMN.AZURE_GPT_4_TURBO_1106
-            | LMN.AZURE_GPT_4_VISION_PREVIEW
             | LMN.AZURE_GPT_4_32K_0613
             | LMN.AZURE_GPT_4_TURBO_2024_0409
         ):
@@ -45,8 +44,10 @@ def get_encoder_name(model_name: LanguageModelName) -> Optional[EncoderName]:
         ):
             return EncoderName.O200K_BASE
         case _:
-            print(f"{model_name} is not supported. Please add encoder information.")
-            return None
+            print(
+                f"{model_name} is not supported. Please add encoder information. Using default"
+            )
+            return EncoderName.CL100K_BASE
 
 
 class LanguageModelProvider(StrEnum):
@@ -54,13 +55,28 @@ class LanguageModelProvider(StrEnum):
     CUSTOM = "CUSTOM"
 
 
+class ModelCapabilities(StrEnum):
+    FUNCTION_CALLING = "function_calling"
+    PARALLEL_FUNCTION_CALLING = "parallel_function_calling"
+    REPRODUCIBLE_OUTPUT = "reproducible_output"
+    STRUCTURED_OUTPUT = "structured_output"
+    VISION = "vision"
+    STREAMING = "streaming"
+    REASONING = "reasoning"
+
+
 class LanguageModelInfo(BaseModel):
     name: LanguageModelName | str
     version: str
     provider: LanguageModelProvider
 
-    encoder_name: Optional[EncoderName] = None
-    token_limits: Optional[LanguageModelTokenLimits] = None
+    encoder_name: EncoderName = EncoderName.CL100K_BASE
+
+    # TODO: Discuss if this is a sensible defaut
+    token_limits: LanguageModelTokenLimits = LanguageModelTokenLimits(
+        token_limit_input=7_000, token_limit_output=1_000
+    )
+    capabilities: list[ModelCapabilities] = [ModelCapabilities.STREAMING]
 
     info_cutoff_at: Optional[date] = None
     published_at: Optional[date] = None
@@ -70,14 +86,20 @@ class LanguageModelInfo(BaseModel):
     retirement_text: Optional[str] = None
 
     @classmethod
-    def from_name(cls, model_name: LanguageModelName):
+    def from_name(cls, model_name: LanguageModelName) -> Self:
         match model_name:
             case LanguageModelName.AZURE_GPT_35_TURBO_0125:
                 return cls(
                     name=model_name,
                     provider=LanguageModelProvider.AZURE,
+                    capabilities=[
+                        ModelCapabilities.STRUCTURED_OUTPUT,
+                        ModelCapabilities.FUNCTION_CALLING,
+                        ModelCapabilities.PARALLEL_FUNCTION_CALLING,
+                        ModelCapabilities.REPRODUCIBLE_OUTPUT,
+                    ],
                     version="0125",
-                    encoder_name=get_encoder_name(model_name),
+                    encoder_name=EncoderName.CL100K_BASE,
                     token_limits=LanguageModelTokenLimits(
                         token_limit_input=16385, token_limit_output=4096
                     ),
@@ -90,43 +112,27 @@ class LanguageModelInfo(BaseModel):
                     name=model_name,
                     provider=LanguageModelProvider.AZURE,
                     version="0613",
-                    encoder_name=get_encoder_name(model_name),
+                    encoder_name=EncoderName.CL100K_BASE,
+                    capabilities=[
+                        ModelCapabilities.FUNCTION_CALLING,
+                        ModelCapabilities.STREAMING,
+                    ],
                     token_limits=LanguageModelTokenLimits(token_limit=8192),
                     info_cutoff_at=date(2021, 9, 1),
                     published_at=date(2023, 6, 13),
                     deprecated_at=date(2024, 10, 1),
                     retirement_at=date(2025, 6, 6),
                 )
-            case LanguageModelName.AZURE_GPT_4_TURBO_1106:
-                return cls(
-                    name=model_name,
-                    provider=LanguageModelProvider.AZURE,
-                    version="1106-preview",
-                    encoder_name=get_encoder_name(model_name),
-                    token_limits=LanguageModelTokenLimits(
-                        token_limit_input=128000, token_limit_output=4096
-                    ),
-                    info_cutoff_at=date(2023, 4, 1),
-                    published_at=date(2023, 11, 6),
-                )
-            case LanguageModelName.AZURE_GPT_4_VISION_PREVIEW:
-                return cls(
-                    name=model_name,
-                    provider=LanguageModelProvider.AZURE,
-                    version="vision-preview",
-                    encoder_name=get_encoder_name(model_name),
-                    token_limits=LanguageModelTokenLimits(
-                        token_limit_input=128000, token_limit_output=4096
-                    ),
-                    info_cutoff_at=date(2023, 4, 1),
-                    published_at=date(2023, 11, 6),
-                )
             case LanguageModelName.AZURE_GPT_4_32K_0613:
                 return cls(
                     name=model_name,
                     provider=LanguageModelProvider.AZURE,
                     version="1106-preview",
-                    encoder_name=get_encoder_name(model_name),
+                    capabilities=[
+                        ModelCapabilities.FUNCTION_CALLING,
+                        ModelCapabilities.STREAMING,
+                    ],
+                    encoder_name=EncoderName.CL100K_BASE,
                     token_limits=LanguageModelTokenLimits(token_limit=32768),
                     info_cutoff_at=date(2021, 9, 1),
                     published_at=date(2023, 6, 13),
@@ -136,7 +142,14 @@ class LanguageModelInfo(BaseModel):
             case LanguageModelName.AZURE_GPT_4_TURBO_2024_0409:
                 return cls(
                     name=model_name,
-                    encoder_name=get_encoder_name(model_name),
+                    encoder_name=EncoderName.CL100K_BASE,
+                    capabilities=[
+                        ModelCapabilities.FUNCTION_CALLING,
+                        ModelCapabilities.PARALLEL_FUNCTION_CALLING,
+                        ModelCapabilities.STRUCTURED_OUTPUT,
+                        ModelCapabilities.VISION,
+                        ModelCapabilities.STREAMING,
+                    ],
                     provider=LanguageModelProvider.AZURE,
                     version="turbo-2024-04-09",
                     token_limits=LanguageModelTokenLimits(
@@ -148,11 +161,18 @@ class LanguageModelInfo(BaseModel):
             case LanguageModelName.AZURE_GPT_4o_2024_0513:
                 return cls(
                     name=model_name,
-                    encoder_name=get_encoder_name(model_name),
+                    encoder_name=EncoderName.O200K_BASE,
+                    capabilities=[
+                        ModelCapabilities.STRUCTURED_OUTPUT,
+                        ModelCapabilities.FUNCTION_CALLING,
+                        ModelCapabilities.PARALLEL_FUNCTION_CALLING,
+                        ModelCapabilities.STREAMING,
+                        ModelCapabilities.VISION,
+                    ],
                     provider=LanguageModelProvider.AZURE,
                     version="2024-05-13",
                     token_limits=LanguageModelTokenLimits(
-                        token_limit_input=128000, token_limit_output=4096
+                        token_limit_input=128_000, token_limit_output=4_096
                     ),
                     info_cutoff_at=date(2023, 10, 1),
                     published_at=date(2024, 5, 13),
@@ -160,11 +180,18 @@ class LanguageModelInfo(BaseModel):
             case LanguageModelName.AZURE_GPT_4o_2024_0806:
                 return cls(
                     name=model_name,
-                    encoder_name=get_encoder_name(model_name),
+                    encoder_name=EncoderName.O200K_BASE,
+                    capabilities=[
+                        ModelCapabilities.STRUCTURED_OUTPUT,
+                        ModelCapabilities.FUNCTION_CALLING,
+                        ModelCapabilities.PARALLEL_FUNCTION_CALLING,
+                        ModelCapabilities.STREAMING,
+                        ModelCapabilities.VISION,
+                    ],
                     provider=LanguageModelProvider.AZURE,
                     version="2024-08-06",
                     token_limits=LanguageModelTokenLimits(
-                        token_limit_input=128000, token_limit_output=16384
+                        token_limit_input=128_000, token_limit_output=16_384
                     ),
                     info_cutoff_at=date(2023, 10, 1),
                     published_at=date(2024, 8, 6),
@@ -172,14 +199,77 @@ class LanguageModelInfo(BaseModel):
             case LanguageModelName.AZURE_GPT_4o_MINI_2024_0718:
                 return cls(
                     name=model_name,
+                    capabilities=[
+                        ModelCapabilities.STRUCTURED_OUTPUT,
+                        ModelCapabilities.FUNCTION_CALLING,
+                        ModelCapabilities.PARALLEL_FUNCTION_CALLING,
+                        ModelCapabilities.STREAMING,
+                        ModelCapabilities.VISION,
+                    ],
                     provider=LanguageModelProvider.AZURE,
                     version="2024-07-18",
-                    encoder_name=get_encoder_name(model_name),
+                    encoder_name=EncoderName.O200K_BASE,
                     token_limits=LanguageModelTokenLimits(
-                        token_limit_input=128000, token_limit_output=16384
+                        token_limit_input=128_000, token_limit_output=16_384
                     ),
                     info_cutoff_at=date(2023, 10, 1),
                     published_at=date(2024, 7, 18),
+                )
+            case LanguageModelName.AZURE_GPT_o1_2024_1217:
+                return cls(
+                    name=model_name,
+                    capabilities=[
+                        ModelCapabilities.STRUCTURED_OUTPUT,
+                        ModelCapabilities.FUNCTION_CALLING,
+                        ModelCapabilities.STREAMING,
+                        ModelCapabilities.VISION,
+                        ModelCapabilities.REASONING,
+                    ],
+                    provider=LanguageModelProvider.AZURE,
+                    version="2024-12-17",
+                    encoder_name=EncoderName.O200K_BASE,
+                    token_limits=LanguageModelTokenLimits(
+                        token_limit_input=200_000, token_limit_output=100_000
+                    ),
+                    info_cutoff_at=date(2023, 10, 1),
+                    published_at=date(2024, 12, 17),
+                )
+            case LanguageModelName.AZURE_GPT_o1_MINI_2024_0912:
+                return cls(
+                    name=model_name,
+                    capabilities=[
+                        ModelCapabilities.STRUCTURED_OUTPUT,
+                        ModelCapabilities.FUNCTION_CALLING,
+                        ModelCapabilities.STREAMING,
+                        ModelCapabilities.VISION,
+                        ModelCapabilities.REASONING,
+                    ],
+                    provider=LanguageModelProvider.AZURE,
+                    version="2024-09-12",
+                    encoder_name=EncoderName.O200K_BASE,
+                    token_limits=LanguageModelTokenLimits(
+                        token_limit_input=128_000, token_limit_output=65_536
+                    ),
+                    info_cutoff_at=date(2023, 10, 1),
+                    published_at=date(2024, 9, 12),
+                )
+            case LanguageModelName.AZURE_GPT_o3_MINI_2025_0131:
+                return cls(
+                    name=model_name,
+                    capabilities=[
+                        ModelCapabilities.STRUCTURED_OUTPUT,
+                        ModelCapabilities.FUNCTION_CALLING,
+                        ModelCapabilities.STREAMING,
+                        ModelCapabilities.REASONING,
+                    ],
+                    provider=LanguageModelProvider.AZURE,
+                    version="2025-01-31",
+                    encoder_name=EncoderName.O200K_BASE,
+                    token_limits=LanguageModelTokenLimits(
+                        token_limit_input=200_000, token_limit_output=100_000
+                    ),
+                    info_cutoff_at=date(2023, 10, 1),
+                    published_at=date(2025, 1, 31),
                 )
             case _:
                 if isinstance(model_name, LanguageModelName):