+++ unique_toolkit/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.6] - 2025-03-10
+- Add o1, o1-mini and o3-mini models
+- Remove deprecated gpt4 models
+- Make token_limits and encoder a required attribute of LanguageModelInfo
+
 ## [0.6.5] - 2025-03-04
 - Add `upload_content_from_bytes` to `ContentService`
 - Add `download_content_to_bytes` to `ContentService`